\name{dgp}
\alias{dgp}
\title{Generates Data from Medeiros and Mendes (2016) ARDL process.}
\usage{
dgp(N,p,T,rho=0.5,betas=rep(1,p-1),innovations="garch",dfg=5,g.par=c(5e-4,0.9,0.05),
             var.error="student.t",dfv=5,size.a.blk=5,a1.par=0.15,a4.par=-0.1)
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}
